requirements.txt
pyyaml == 3.13
python-gitlab == 1.7.0